Network클래스 추출, getCardAvailable Card가 아닌 CardPayment로 옮김
